Revert "refactor: rename resource files."
