+++ src/index.ts
@@ -1,17 +1,17 @@
-import Accounts from './accounts';
-import Charges from './charges';
-import Customers from './customers';
-import Events from './events';
-import Plans from './plans';
-import Subscriptions from './subscriptions';
+import Accounts from './account';
+import Charges from './charge';
+import Customers from './customer';
+import Events from './event';
+import Plans from './plan';
+import Subscriptions from './subscription';
 import Tenants from './tenants';
 import TenantTransfers from './tenantTransfers';
-import Tokens from './tokens';
-import Transfers from './transfers';
-import Statements from './statements';
-import Terms from "./terms";
-import Balances from "./balances";
-import ThreeDSecureRequests from "./threeDSecureRequests";
+import Tokens from './token';
+import Transfers from './transfer';
+import Statements from './statement';
+import Terms from "./term";
+import Balances from "./balance";
+import ThreeDSecureRequests from "./threeDSecureRequest";
 
 namespace Payjp {
   export interface PayjpStatic {
