+++ src/term.ts
