+++ src/token.ts
