+++ test/balance.spec.js
