+++ test/customer.spec.js
