+++ test/statement.spec.js
