+++ test/term.spec.js
