+++ test/token.spec.js
